// src/vs/platform/agentHost/node/codex/protocol/generated/v2/FileChangePatchUpdatedNotification.ts
// AUTOGENERATED - do not edit.
// Source: `codex app-server generate-ts --experimental`
// Generated from @openai/codex 0.134.0, licensed Apache-2.0.
// Modified by VS Code to rewrite imports and apply repository formatting.
// See README.md in this directory for provenance and licensing notes.
// Regenerate: npm run codex:gen-protocol
// Source file: v2/FileChangePatchUpdatedNotification.ts

import type { FileUpdateChange } from "./FileUpdateChange.js";

export type FileChangePatchUpdatedNotification = { threadId: string, turnId: string, itemId: string, changes: Array<FileUpdateChange>, };